requirements.txt
facenet_pytorch==2.5.3
fastapi==0.96.1
numpy==1.21.5
Pillow==9.5.0
pydantic==1.10.9
requests==2.28.1
torch==2.0.1
uvicorn==0.22.0
~pencv_python_headless==4.5.5.62
